requirements.txt
Pillow==9.4.0
pygame==2.1.2
